# src/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ensemble",
    "description": "AI-çağı ekipleri için paylaşılan proje beyni",
    "version": "0.0.1"
  },
  "paths": {
...
    "/presence": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Presence",
        "description": "Aktif çalışan beyanlarını döndürür; bayat kayıtlar read-time'da elenir (#60).",
        "operationId": "get_presence_presence_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponse"
                }
              }
            }
          },
          "502": {
            "description": "Kalici saglayici hatasi (GitHub/Gemini/Ollama)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Gecici olarak erisilemez",
            "headers": {
              "Retry-After": {
                "description": "Saniye — yalniz kendiliginden duzelebilir durumlarda",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Demo istek limiti aşıldı (yalnız DEMO_MODE)",
            "headers": {
              "Retry-After": {
                "description": "Saniye — pencere kayınca tekrar denenebilir",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "?repo= izinli sette değil (T-79)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Events",
        "description": "GitHub event akışını cursor'lı döndürür (#52).\n\n- `since`: bir önceki yanıtın `latest_ts`'i; payload'ı yeni eventlerle sınırlar.\n- `If-None-Match`: bir önceki ETag; içerik değişmediyse gövde yerine 304 döner.",
        "operationId": "get_events_events_get",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artımlı polling cursor. Verilirse yalnız bu andan (dahil) sonraki eventler döner.",
              "title": "Since"
            },
            "description": "Artımlı polling cursor. Verilirse yalnız bu andan (dahil) sonraki eventler döner."
          },
          {
            "name": "if-none-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "502": {
            "description": "Kalici saglayici hatasi (GitHub/Gemini/Ollama)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Gecici olarak erisilemez",
            "headers": {
              "Retry-After": {
                "description": "Saniye — yalniz kendiliginden duzelebilir durumlarda",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Demo istek limiti aşıldı (yalnız DEMO_MODE)",
            "headers": {
              "Retry-After": {
                "description": "Saniye — pencere kayınca tekrar denenebilir",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "?repo= izinli sette değil (T-79)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/github": {
      "post": {
        "tags": [
          "webhook"
        ],
        "summary": "Github Webhook",
        "operationId": "github_webhook_webhooks_github_post",
        "parameters": [
          {
            "name": "x-hub-signature-256",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hub-Signature-256"
            }
          },
          {
            "name": "x-github-event",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Github-Event"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Github Webhook Webhooks Github Post"
                }
              }
            }
          },
          "502": {
            "description": "Kalici saglayici hatasi (GitHub/Gemini/Ollama)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Gecici olarak erisilemez",
            "headers": {
              "Retry-After": {
                "description": "Saniye — yalniz kendiliginden duzelebilir durumlarda",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Demo istek limiti aşıldı (yalnız DEMO_MODE)",
            "headers": {
              "Retry-After": {
                "description": "Saniye — pencere kayınca tekrar denenebilir",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz JSON gövdesi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Eksik/geçersiz webhook imzası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorRef": {
        "properties": {
          "handle": {
            "type": "string",
            "title": "Handle"
          },
          "type": {
            "type": "string",
            "enum": [
              "human",
              "agent"
            ],
            "title": "Type"
          },
          "responsible": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible"
          }
        },
        "type": "object",
        "required": [
          "handle",
          "type"
        ],
        "title": "ActorRef",
        "description": "Açık aktör tipi (#32/#52) — kontrat: docs/sprint2-kontratlar.md Ek B1."
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "status"
        ],
        "title": "ErrorEnvelope"
      },
      "EventsResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/NormalizedEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "latest_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Latest Ts"
          }
        },
        "type": "object",
        "required": [
          "events",
          "latest_ts"
        ],
        "title": "EventsResponse",
        "description": "GET /events yanıt modeli.\n\nlatest_ts, bir sonraki poll'de `since` sorgu parametresi olarak geri\ngönderilecek cursor değeridir (#52)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NormalizedEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "commit",
              "pr",
              "issue",
              "branch"
            ],
            "title": "Type"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "actor",
          "branch",
          "files",
          "ts",
          "ref"
        ],
        "title": "NormalizedEvent"
      },
      "PresenceEntry": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/ActorRef"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "since": {
            "type": "string",
            "format": "date-time",
            "title": "Since"
          }
        },
        "type": "object",
        "required": [
          "actor",
          "module",
          "task",
          "branch",
          "since"
        ],
        "title": "PresenceEntry",
        "description": ".harness/active/* projeksiyonu (#32/#52) — kontrat: Ek B1."
      },
      "PresenceResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/PresenceEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "latest_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Latest Ts"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "latest_ts"
        ],
        "title": "PresenceResponse",
        "description": "GET /presence yanıt modeli."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}